Migrate the Sonnet examples to the fixed framework
- Drop every inference.py's dead `_warm` monkeypatch (staging warms from
  sample_request.json now).
- 03/05/06/11: type the date-carrying reads of their frame steps
  (`reads={"accounts": list[Account], ...}`) so served JSON dates are coerced.
- One type per input column: annotate as float the steps that read a column a
  decision table or scorecard reads as float (product_code, segment_code,
  term_months, employment_type_code, months_on_book, ...); 10 gives its buffer
  table a float copy of channel_code and drops a relabel of a read that step
  never had.
- Serve 00 and 02-11 in fused mode (output equals interpreted on each
  sample_request.json); 01 stays interpreted (fused tree compile hits a
  RecursionError in the tree walker).

+++ example_projects/00-shared-credit-core/sonnet/credit_core/adjustments.py
@@ -218,7 +218,7 @@ class AdjustmentRegister:
         def apply_adjustments(
             adjustment_base_value: float, decision_date: date,
             adjustment_stack_enabled: bool = param(True),
-            product_code: int = missing_as(-1), segment_code: int = missing_as(-1),
+            product_code: float = missing_as(-1.0), segment_code: float = missing_as(-1.0),
             channel_code: int = missing_as(-1), scorecard_id: int = missing_as(-1),
         ) -> tuple[float, float, str, list[str]]:
             provenance = {"product_code": product_code, "segment_code": segment_code,
